requirements.txt
asgiref==3.2.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-resized==0.3.11
django-storages==1.9.1
djangorestframework==3.11.0
google-api-core==1.21.0
google-auth==1.19.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
